requirements.txt
Flask
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
dateparser
google-generativeai
